--feat=add text_mode
